Simplify FlightTicket, since interface/implementation splitting is no advantage in this particular case (data beans) and now is easier to read and maintain code.
